// apps/arabic-mandi/src/app/Core/admin/generate-admin-types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type BooleanFieldComparison = {
  is?: InputMaybe<Scalars['Boolean']>;
  isNot?: InputMaybe<Scalars['Boolean']>;
};
// ...
export type FoodCategory = {
  __typename?: 'FoodCategory';
  category_image?: Maybe<Scalars['String']>;
  clientView: Scalars['Boolean'];
  id: Scalars['Float'];
  isActive: Scalars['Boolean'];
  items?: Maybe<ItemEntity>;
  name: Scalars['String'];
};
// ...
export type FoodCategoryFilter = {
  and?: InputMaybe<Array<FoodCategoryFilter>>;
  clientView?: InputMaybe<BooleanFieldComparison>;
  id?: InputMaybe<NumberFieldComparison>;
  isActive?: InputMaybe<BooleanFieldComparison>;
  items?: InputMaybe<FoodCategoryFilterItemEntityFilter>;
  name?: InputMaybe<StringFieldComparison>;
  or?: InputMaybe<Array<FoodCategoryFilter>>;
};

export type FoodCategoryFilterItemEntityFilter = {
  and?: InputMaybe<Array<FoodCategoryFilterItemEntityFilter>>;
  id?: InputMaybe<NumberFieldComparison>;
  name?: InputMaybe<StringFieldComparison>;
  offer?: InputMaybe<StringFieldComparison>;
  or?: InputMaybe<Array<FoodCategoryFilterItemEntityFilter>>;
  status?: InputMaybe<BooleanFieldComparison>;
  type?: InputMaybe<BooleanFieldComparison>;
};

export type FoodCategoryInput = {
  category_image?: InputMaybe<Scalars['String']>;
  clientView?: InputMaybe<Scalars['Boolean']>;
  id?: InputMaybe<Scalars['Float']>;
  isActive?: InputMaybe<Scalars['Boolean']>;
  name?: InputMaybe<Scalars['String']>;
};

export type FoodCategoryMaxAggregate = {
  __typename?: 'FoodCategoryMaxAggregate';
  id?: Maybe<Scalars['Float']>;
  name?: Maybe<Scalars['String']>;
};

export type FoodCategoryMinAggregate = {
  __typename?: 'FoodCategoryMinAggregate';
  id?: Maybe<Scalars['Float']>;
  name?: Maybe<Scalars['String']>;
};

export type FoodCategorySort = {
  direction: SortDirection;
  field: FoodCategorySortFields;
  nulls?: InputMaybe<SortNulls>;
};

export enum FoodCategorySortFields {
  ClientView = 'clientView',
  Id = 'id',
  IsActive = 'isActive',
  Name = 'name'
}

export type FoodCategorySumAggregate = {
  __typename?: 'FoodCategorySumAggregate';
  id?: Maybe<Scalars['Float']>;
};

export type FoodSize = {
  __typename?: 'FoodSize';
  id: Scalars['Float'];
  isActive: Scalars['Boolean'];
  name: Scalars['String'];
};
// ...
export type FoodSizeFilter = {
  and?: InputMaybe<Array<FoodSizeFilter>>;
  id?: InputMaybe<NumberFieldComparison>;
  isActive?: InputMaybe<BooleanFieldComparison>;
  name?: InputMaybe<StringFieldComparison>;
  or?: InputMaybe<Array<FoodSizeFilter>>;
};

export type FoodSizeInput = {
  id?: InputMaybe<Scalars['Float']>;
  name?: InputMaybe<Scalars['String']>;
};

export type FoodSizeMaxAggregate = {
  __typename?: 'FoodSizeMaxAggregate';
  id?: Maybe<Scalars['Float']>;
  name?: Maybe<Scalars['String']>;
};

export type FoodSizeMinAggregate = {
  __typename?: 'FoodSizeMinAggregate';
  id?: Maybe<Scalars['Float']>;
  name?: Maybe<Scalars['String']>;
};

export type FoodSizeSort = {
  direction: SortDirection;
  field: FoodSizeSortFields;
  nulls?: InputMaybe<SortNulls>;
};

export enum FoodSizeSortFields {
  Id = 'id',
  IsActive = 'isActive',
  Name = 'name'
}

export type FoodSizeSumAggregate = {
  __typename?: 'FoodSizeSumAggregate';
  id?: Maybe<Scalars['Float']>;
};

export type FoodType = {
  __typename?: 'FoodType';
  id: Scalars['Float'];
  isActive: Scalars['Boolean'];
  name: Scalars['String'];
};
// ...
export type FoodTypeFilter = {
  and?: InputMaybe<Array<FoodTypeFilter>>;
  id?: InputMaybe<NumberFieldComparison>;
  isActive?: InputMaybe<BooleanFieldComparison>;
  name?: InputMaybe<StringFieldComparison>;
  or?: InputMaybe<Array<FoodTypeFilter>>;
};

export type FoodTypeInput = {
  id?: InputMaybe<Scalars['Float']>;
  name?: InputMaybe<Scalars['String']>;
};

export type FoodTypeMaxAggregate = {
  __typename?: 'FoodTypeMaxAggregate';
  id?: Maybe<Scalars['Float']>;
  name?: Maybe<Scalars['String']>;
};

export type FoodTypeMinAggregate = {
  __typename?: 'FoodTypeMinAggregate';
  id?: Maybe<Scalars['Float']>;
  name?: Maybe<Scalars['String']>;
};

export type FoodTypeSort = {
  direction: SortDirection;
  field: FoodTypeSortFields;
  nulls?: InputMaybe<SortNulls>;
};

export enum FoodTypeSortFields {
  Id = 'id',
  IsActive = 'isActive',
  Name = 'name'
}

export type FoodTypeSumAggregate = {
  __typename?: 'FoodTypeSumAggregate';
  id?: Maybe<Scalars['Float']>;
};

export type ItemEntity = {
  __typename?: 'ItemEntity';
  category?: Maybe<FoodCategory>;
  createdby?: Maybe<User>;
  foodsize?: Maybe<FoodSize>;
  foodtype?: Maybe<FoodType>;
  id: Scalars['Float'];
  image_data?: Maybe<Scalars['String']>;
  name: Scalars['String'];
  offer: Scalars['String'];
  price: Scalars['Float'];
  status: Scalars['Boolean'];
  type: Scalars['Boolean'];
  updatedby?: Maybe<User>;
};
// ...
export type NumberFieldComparison = {
  between?: InputMaybe<NumberFieldComparisonBetween>;
  eq?: InputMaybe<Scalars['Float']>;
  gt?: InputMaybe<Scalars['Float']>;
  gte?: InputMaybe<Scalars['Float']>;
  in?: InputMaybe<Array<Scalars['Float']>>;
  is?: InputMaybe<Scalars['Boolean']>;
  isNot?: InputMaybe<Scalars['Boolean']>;
  lt?: InputMaybe<Scalars['Float']>;
  lte?: InputMaybe<Scalars['Float']>;
  neq?: InputMaybe<Scalars['Float']>;
  notBetween?: InputMaybe<NumberFieldComparisonBetween>;
  notIn?: InputMaybe<Array<Scalars['Float']>>;
};

export type NumberFieldComparisonBetween = {
  lower: Scalars['Float'];
  upper: Scalars['Float'];
};

export type Offer = {
  __typename?: 'Offer';
  createdby?: Maybe<User>;
  discount: Scalars['Float'];
  id: Scalars['ID'];
  items?: Maybe<Array<OfferItem>>;
  name: Scalars['String'];
  price: Scalars['Float'];
  status: Scalars['Boolean'];
  totalPrice: Scalars['Float'];
  updatedby?: Maybe<User>;
};
// ...
export type Query = {
  __typename?: 'Query';
  foodCategories: Array<FoodCategory>;
  foodCategory?: Maybe<FoodCategory>;
  foodSize?: Maybe<FoodSize>;
  foodSizes: Array<FoodSize>;
  foodType?: Maybe<FoodType>;
  foodTypes: Array<FoodType>;
  getItems: Array<ItemEntity>;
  itemEntities: Array<ItemEntity>;
  itemEntity?: Maybe<ItemEntity>;
  offer?: Maybe<Offer>;
  offers: Array<Offer>;
  user?: Maybe<User>;
  users: Array<User>;
};
// ...
export enum SortDirection {
  Asc = 'ASC',
  Desc = 'DESC'
}

/** Sort Nulls Options */
export enum SortNulls {
  NullsFirst = 'NULLS_FIRST',
  NullsLast = 'NULLS_LAST'
}

export type StringFieldComparison = {
  eq?: InputMaybe<Scalars['String']>;
  gt?: InputMaybe<Scalars['String']>;
  gte?: InputMaybe<Scalars['String']>;
  iLike?: InputMaybe<Scalars['String']>;
  in?: InputMaybe<Array<Scalars['String']>>;
  is?: InputMaybe<Scalars['Boolean']>;
  isNot?: InputMaybe<Scalars['Boolean']>;
  like?: InputMaybe<Scalars['String']>;
  lt?: InputMaybe<Scalars['String']>;
  lte?: InputMaybe<Scalars['String']>;
  neq?: InputMaybe<Scalars['String']>;
  notILike?: InputMaybe<Scalars['String']>;
  notIn?: InputMaybe<Array<Scalars['String']>>;
  notLike?: InputMaybe<Scalars['String']>;
};
// ...
export type UpdateItemInput = {
  category?: InputMaybe<FoodCategoryInput>;
  foodsize?: InputMaybe<FoodSizeInput>;
  foodtype?: InputMaybe<FoodTypeInput>;
  image_data?: InputMaybe<Scalars['String']>;
  name?: InputMaybe<Scalars['String']>;
  offer?: InputMaybe<Scalars['String']>;
  price?: InputMaybe<Scalars['Float']>;
  status?: InputMaybe<Scalars['Boolean']>;
  type?: InputMaybe<Scalars['Boolean']>;
  updatedby?: InputMaybe<UserInput>;
};
// ...
export type UpdateOneItemEntityInput = {
  /** The id of the record to update */
  id: Scalars['ID'];
  /** The update to apply. */
  update: UpdateItemInput;
};
// ...
export type User = {
  __typename?: 'User';
  id: Scalars['Float'];
  name?: Maybe<Scalars['String']>;
  profile_img?: Maybe<Scalars['String']>;
  role: Scalars['String'];
  status: Scalars['Boolean'];
};
// ...
export type UserInput = {
  createdBy?: InputMaybe<Scalars['String']>;
  id?: InputMaybe<Scalars['Float']>;
  name?: InputMaybe<Scalars['String']>;
  profile_img?: InputMaybe<Scalars['String']>;
  role?: InputMaybe<Scalars['String']>;
  updatedBy?: InputMaybe<Scalars['String']>;
};
// ...
export type GetFoodCategoriesQueryVariables = Exact<{
  filter?: InputMaybe<FoodCategoryFilter>;
  sorting?: InputMaybe<Array<FoodCategorySort> | FoodCategorySort>;
}>;
// ...
export type FoodTypesQueryVariables = Exact<{
  filter: FoodTypeFilter;
  sorting?: InputMaybe<Array<FoodTypeSort> | FoodTypeSort>;
}>;


export type FoodTypesQuery = { __typename?: 'Query', foodTypes: Array<{ __typename?: 'FoodType', id: number, name: string }> };

export type FoodSizesQueryVariables = Exact<{
  filter: FoodSizeFilter;
  sorting?: InputMaybe<Array<FoodSizeSort> | FoodSizeSort>;
}>;


export type FoodSizesQuery = { __typename?: 'Query', foodSizes: Array<{ __typename?: 'FoodSize', id: number, name: string }> };